+++ src/test/java/Calculator/CalculatorTest.java
@@ -2,8 +2,11 @@ package Calculator;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.NullAndEmptySource;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 class CalculatorTest {
 
@@ -20,62 +23,11 @@ class CalculatorTest {
         int expected = 10;
         assertEquals(expected, actual);
     }
-
-    @Test
-    @DisplayName("덧셈")
-    void plus() {
+    
+    @ParameterizedTest(name = "\"{0}\" 입력 시 IllegalArgumentException throw")
+    @NullAndEmptySource
+    void inputIsNulloreEmpty(String inputString) {
         //given //when
-        Calculator calculator = new Calculator();
-        int actual = calculator.add(3, 5);
-
-        //then
-        int expected = 8;
-        assertEquals(expected, actual);
-    }
-
-    @Test
-    @DisplayName("뺄셈")
-    void minus() {
-        //given //when
-        Calculator calculator = new Calculator();
-        int actual = calculator.minus(3, 5);
-
-        //then
-        int expected = -2;
-        assertEquals(expected, actual);
-    }
-
-    @Test
-    @DisplayName("곱셈")
-    void multiply() {
-        //given //when
-        Calculator calculator = new Calculator();
-        int actual = calculator.multiply(3, 5);
-
-        //then
-        int expected = 15;
-        assertEquals(expected, actual);
-    }
-
-    @Test
-    @DisplayName("나눗셈")
-    void divide() {
-        //given //when
-        Calculator calculator = new Calculator();
-        int actual = calculator.divide(4, 2);
-
-        //then
-        int expected = 2;
-        assertEquals(expected, actual);
-    }
-
-    @Test
-    @DisplayName("입력 값이 null이거나 빈 공백 문자일 경우 IllegalArgumentException throw")
-    void inputIsNulloreEmpty() {
-        //given
-        String inputString = "";
-
-        //when
         Exception actual = assertThrows(IllegalArgumentException.class, () -> {
             Calculator calculator = new Calculator(inputString);
         });
@@ -84,19 +36,4 @@ class CalculatorTest {
         assertEquals("입력 값이 null이거나 빈 공백 문자입니다.", actual.getMessage());
     }
 
-    @Test
-    @DisplayName("사칙연산 기호가 아닌 경우 IllegalArgumentException throw")
-    void notOperator() {
-        //given
-        String operator = "$";
-        Calculator calculator = new Calculator();
-        
-        //when
-        Exception actual = assertThrows(IllegalArgumentException.class, () -> {
-            calculator.calculateProcess(3, 5, operator);
-        });
-
-        //then
-        assertEquals(operator + "는 사칙연산 기호가 아닙니다: ", actual.getMessage());
-    }
 }
